requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.25
PyMySQL==1.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
